requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
numpy==1.16.0
phue==1.0
Pillow==5.4.1
pprint==0.1
python-dotenv==0.10.1
requests==2.21.0
scipy==1.2.0
six==1.12.0
spotipy==2.4.4
urllib3==1.24.1
